requirements.txt
aiohttp
async-timeout
asyncio-throttle
beautifulsoup4
click
chromadb
crawl4ai
fastapi
groq
httpx
langchain-text-splitters
langdetect
loguru
lxml
numpy
prometheus-fastapi-instrumentator
psutil
pydantic
pydantic-settings
python-dotenv
python-multipart
pyspellchecker
pyyaml
rank-bm25
rich
sentence-transformers
simhash
tqdm
uvicorn

